Reimplemented PictureLoader as a singleton
 * Removed remaining picture handling from card database and cardinfo
 * removed the “noCard object” hack
