+++ cockatrice/src/cardframe.cpp
@@ -93,7 +93,8 @@ void CardFrame::setCard(CardInfo *card)
     if (info)
         disconnect(info, 0, this, 0);
     info = card;
-    connect(info, SIGNAL(destroyed()), this, SLOT(clear()));
+    if(info)
+        connect(info, SIGNAL(destroyed()), this, SLOT(clear()));
     text->setCard(info);
     pic->setCard(info);
 }
