+++ cockatrice/src/dlg_settings.h
@@ -45,6 +45,8 @@ private slots:
     void tokenDatabasePathButtonClicked();
     void languageBoxChanged(int index);
     void setEnabledStatus(bool);
+    void defaultUrlRestoreButtonClicked();
+    void fallbackUrlRestoreButtonClicked();
 private:
     QStringList findQmFiles();
     QString languageName(const QString &qmFile);
@@ -53,13 +55,13 @@ private:
     QLineEdit *picsPathEdit;
     QLineEdit *cardDatabasePathEdit;
     QLineEdit *tokenDatabasePathEdit;
-    QLineEdit *highQualityURLEdit;
+    QLineEdit *defaultUrlEdit;
+    QLineEdit *fallbackUrlEdit;
     QSpinBox pixmapCacheEdit;
     QGroupBox *personalGroupBox; 
     QGroupBox *pathsGroupBox;
     QComboBox languageBox;
     QCheckBox picDownloadCheckBox;
-    QCheckBox picDownloadHqCheckBox;
     QCheckBox updateNotificationCheckBox;
     QLabel languageLabel;
     QLabel pixmapCacheLabel;
@@ -68,9 +70,12 @@ private:
     QLabel picsPathLabel;
     QLabel cardDatabasePathLabel;
     QLabel tokenDatabasePathLabel;
-    QLabel highQualityURLLabel;
-    QLabel highQualityURLLinkLabel;
+    QLabel defaultUrlLabel;
+    QLabel fallbackUrlLabel;
+    QLabel urlLinkLabel;
     QPushButton clearDownloadedPicsButton;
+    QPushButton defaultUrlRestoreButton;
+    QPushButton fallbackUrlRestoreButton;
 };
 
 class AppearanceSettingsPage : public AbstractSettingsPage {
